iegen/python_only:
Changed Inequality to represent GTE rather than LTE.
Canonicalized Equality and Inequality to to only have a single expression: either exp>=0 or exp=0.
Noted that != is not parsed correctly but did not implement a fix, only throws an exception.
Implemented a visitor to translate a PresSet to a constraint matrix that CLooG can use.