Start of iegen script for running IEGen
Added a bin/ directory that contains the Python script 'iegen'.  This
script will be the 'entry point' to IEGen.  Rather than running a spec
file directly, this script will be run and the spec file will be
specified on the command line.  This allows settings such as debug mode
to be enabled or disabled.

References #23.