Updates to IEGenObject, settings and printing
The settings field of IEGenObject is now a class member (static) rather
than an instance member.  This is necessary as a new IEGenSettings
object cannot be created for each instance of IEGenObject.

Added two versions of info_print: instance and static versions.  This
method should be used rather than just using the print statement when
information needs to be printed.

Added a static version of debug_print.

References #27.