Printing updates and iegen script final changes
Added option to optionally print code, the default is to not print the
generated code.

Changed default for output file to None.  The code is not written to a
file by default.  Changed MapIR.codegen() to accomodate this change.

Updated all print statements to use IEGenObject.print_info so that this
output can optionally be turned on and off.

Closes #23.