Added CollectVarsVisitor
While working on issue #43, the issue with executor codegen (bug #37)
was discovered.  This commit adds a new visitor, CollectVarsVisitor.
This visitor collects all tuple variables in a formula.  Additionally, a
function, variables(), was added to Formula that uses this visitor.
Tests were upated for these changes.

This commit will be used for fixing bug #37.