Simplification registration/callback system
This commit closes #52, implenting the simplification rule
registration/callback system.  The code for the simplification rule
functions were moved to a new file.  The registration related code was
implemented.  The loop that iterates over rule functions was generalized
to use the registered rules.

A small performance benefit may result from this commit as the
SortVisitor was previously being called every iteration of every loop.
It turns out this is not needed.  I believe it is many small fixes like
this that will allow performance to be improved.