Added UniqueTupleVarsVisitor, called from parser
UniqueTupleVarsVisitor looks for cases in Sets/Relations where tuple
variable names are duplicated.  If any are found, the variables are
renamed to be unique and equality constraints are added.

This visitor is called once after a Set/Relation is parsed so that any
duplicate names are removed before the Set/Relation can be used.  This
closes bug #77.