Added registration of function inverse pairs
Per feature #80, the ability to register pairs of functions that are
inverses of each other was added.

Added function to register a pair of function names that are inverses of
each other.

The inverse simplification rule function was moved to the file with the
other rules.  This rule is always registered now rather than only being
registered when a MapIR instance is created.  This rule now uses the
names of inverse function pairs rather than obtained from
iegen.simplify.inverse_pairs() rather than obtaining these from a MapIR
instance.