Fix for bug #89
This commit closes #89.  RemoveFreeVarFunctionVisitor was fixed so that
it does not die when certain AST nodes in certain configurations were
given.  A test was added to ensure that the bug was fixed.  The
inverse_simplify rule was upated to not check for certain types to avoid
this problem.  The check was added rather than fixing it as it should
have been.