Fixed slow RelationTestCase.testArity
Closes #86: RelationTestCase.testArity was running slow after the recent
addition of UniqueTupleVarVisitor and the associated simplification
rule.  This was due to the fact that the relations being constructed
(100 in total) had duplicate tuple variable names.  This means that the
UniqueTupleVarVisitor rule was being triggered and constraints were
being added.  When more constraints are present, things run more slowly.
When doing this 100 times, the slowdown was quite apparent.