Update access relations to full iteration space
Closes #38: Access relations are specified as relations from iterators
to a data space.  However, our algorithms assume that access relations
are in terms of the computation's full iteration space.  Therefore this
commit updates the access relations' iter_to_data fields to be in terms
of the full iteration space rather than just the iterators.

Changing the access relations broke the implementation of
DataPermuteTrans.calc_input.  This method was updated to work with the
new format of the access relations.