Fixed problem due to unrename, makes #104 go away
The issue was a naming conflict in the input tuple
vars of two relations we were attempting to compose.
The unrename caused two different variables to have the
same name.  Now we just don't unrename if there are any
variable conflicts between the two relations.
Also made the compose debug output prettier.