Added implementation of versioned data arrays
References #114: the VersionedDataArray class was added.  This is a
wrapper class for data arrays that represents a specific version of the
data array at a given time in the IDG.  The calc_initial_idg function
was updated to use this class.  The DotVisitor was updated to print the
version number of the data array in the IDG.  Finally, DataPermuteTrans
was updated to use the VersionedDataArray class.