Added OutputERVisitor
References #100: In the process of debugging C drivers for the simple
examples, I discovered that the output ER specs were not being assigned
to a local _ER variable.  The OutputERVisitor was added to remedy this
problem.  This visitor adds a declaration of the form 'name_ER=*name;'
for each output ER spec.

A call to this visitor was added to the gen_executor function.