IDG dependence building in main trans loop
References #113: This commit moves the calculation and creation of IDG
dependences related to ERSpecs from DataPermuteTrans to the main
transformation application loop as an additional step each iteration.
This greatly reduces the amount of code needed in a transformation and
leaves the transformation write to worry about nodes and dependences
specific to the transformation rather than the general dependences on
symbolics, ERs, etc.