Create scattering dimension names in pycloog
References #133: This commit fixes a bug in the pycloog C code.  When
scattering functions were specified, no names were created for the
dimensions of the scattering functions.  This was causing the
cloog_program_pprint function to die.  This commit creates these names
if necessary.  Also, if the number of iterators is the same as the
number of non-scalar scattering dimensions, the iterator names are used
for the non-scalar scattering dimensions.  This allows the same iterator
names to be preserved.