Define context for parameters in pycloog
References #133: the pycloog_get_context function now by default defines
the context for all parameters to be >=0.  This is a safe assumption for
all of the usages in the forseeable future.  This may need to be changed
at some point.  However, the conversion to using the bcloog project
rather than the integrated pycloog CLooG bindings may resolve this
issue.