Updated TransVisitor to translate relations
References #133: The TransVisitor now translates a given Set into a
union of CLooG matrix domains and translates a given Relation with a
single conjunction into a scattering function in CLooG matrix format.
Tests were updated and added to ensure this visitor is still working
properly for Sets and that it works properly for Relations.