Implemented ability to include other spec files
References #119: This commit adds the iegen.include function.  This
function can be used from within a spec file to include another spec
file.  This allows a computation to be specified in a single spec file
and other spec files that apply various transformations to this spec
file to be written without duplication the original computation
specification.