Implemented a faster version of FOREACH_in_tuple
The implementation I initially used iterated over the size of the
rectangular domain and generated the full tuple at each point with
the function ER_calcTuple.

Timings with CFLAGS=-O3 -DNDEBUG on the test_ExplicitRelation driver with
K=30 in the stress test were as follows:

==== Done checking the {[x,y]->[y,x]} stress test
12.166u 2.882s 0:15.10 99.6%    0+0k 0+0io 0pf+0w
==== Done checking the {[x,y]->[y,x]} stress test
12.176u 2.910s 0:15.11 99.8%    0+0k 0+0io 0pf+0w

A faster implementation will attempt to increment the last element in
the tuple until it hits its upper bound and then increment the preceding
elements accordingly to induce a lexicographical order.

Timings with the new implementation of the FOREACH iterator and with
CFLAGS=-O3 -DNDEBUG on the test_ExplicitRelation driver with K=30 in
the stress test were as follows:

==== Done checking the {[x,y]->[y,x]} stress test
7.131u 1.677s 0:08.82 99.7% 0+0k 0+0io 0pf+0w
==== Done checking the {[x,y]->[y,x]} stress test
7.114u 1.689s 0:08.86 99.2% 0+0k 0+0io 0pf+0w

The new iterator is almost twice as fast.  Cool!