More updates to user specified types
Users now have the option of specifying a 'hole' in the type string for
a data array, index array, or symbolic.  For example, the user can
specify 'double %s[][9]' for a data array type.  If a hole is present,
it is filled with the name of the data array.  This was added to support
interleaving of data arrays.

References #195