Added ability to specify symbolic bounds
Upper and lower bounds can now be specified for symbolics.  If
specified, these bounds are passed to the code generator and
consequently on to CLooG as the context for the parameters when CLooG
generates code.  The net effect of this change is that the user can now
specify bounds to get rid of guards that CLooG generates when no bounds
are specified for parameters.

The example spec files were updated to specify lower bounds of 1 for
each declared symbolic.

Closes #208