First part of new data structure is complete
The first portion of implementing the new data structure is now working
and has associated tests.  This portion takes the specified formula
string, parses it, and sets up the tuple variable, symbolic, and free
variable names and associates these names with column positions.  The
next step is to translate the parsed constrants from AST form to the new
data strcture format.  This will be done using one or more visitors (see
issue #210).

Closes #209.