Added inverse simplification callback facility
This commit adds the ability to register a callback with the
simplification system to be notified when the inverse simplification
rule fires.  This is needed to be able to know when to add an ERSpec for
an inverse function.  This commit lays the foundation to be able to
complete issue #102.

References #102,#113