Fixed bug with IterPermuteTrans
It turns out the bug related to IterPermuteTrans had something to do
with naming (probably in compose).  I simply renamed a tuple variable
and the problem went away.  May need to track down the actual cause of
this rather than just using the work-around, but for now this is fixed.

Closes #84