Inverse ERSpecs are now added to the IDG
The MapIR is notified any time the inverse simplification rule files.
The handler for this notification now creates an ERSpec for the inverse
of the function that was the target of the simplification rule.  The
appropriate nodes and dependences are added to the IDG.  Code generation
was updated to generate inverse ERs.

Closes #102