Sparse formula constraint support (no functions)
This commit adds support for affine constraints to sparse formulas.
Functions are not yet supported.  The SparseTransVisitor was updated to
create coefficient vectors based on the AST that it visits.  These
coefficient vectors are added as equalities and inequalities to a sparse
formula.  Many of the sparse formula supporting classes were implemented
and fleshed out.

Still to do:  Implement formula support and equality for sparse
formulas.

References #209,#210