Added partial UFCall implementation
The SparseTransVisitor now adds functions to the formula AST it is
visiting.  The UFCall class is mostly implemented at this point.
However, I've run into an issue related to ordering of functions.  I
believe I may need to rethink how we are representing coefficients as I
don't think we want to attempt to impose an ordering on UFCall
instances.

References #209,#210