Sparse Set/Relation construction completed
The SparseTransVisitor now supports functions.  This means that the
set/relations strings that Sparse Set/Relation support are equivalent to
those of the current Set/Relation implementation.

This commit changes the way the coefficients are represented.  They are
now completely unordered and are stored as dictionaries rather than
ordered lists or tuples.

Closes #210