Refactoring Sparse{Formula,Set,Relation} classes
The interface to the SparseFormula constructor has changed to accept
different arguments now.  Code in that constructor has moved to
individual methods.  The SparseSet and SparseRelation constructors now
use this changed constructor interface for SparseFormula.  This is the
start of changes that will allow the set/relation operations to be
implemented in a clean way.

References #233