Added SparseDisjunction and SparseConjunction
This commit adds the SparseDisjunction and SparseDisjunction classes.
SparseDisjunction represents a disjunction of a collection of
conjunctions.  SparseConjunction represents the conjunction of a
collection of equality and inequality constraints.  These classes were
added to make managing the data they represent easier.  Previously it
was represented as a 'set of sets'.  Now there are classes with methods
for copying and printing.  This makes the code that uses these classes
much cleaner.

References #233