SparseRelation.inverse bug fix
This commit fixes two bugs with the SparseRelation.inverse operation.
First, the mappings of old position to new position for the tuple
variables wasn't being calculated correctly, causing an exception to be
raised.  Second, the input arity for the new inverted relation wasn't
being set, and therefore the result was incorrect.

References #233