Apply/Compose operations renaming fix
This commit fixes an issue with the apply and compose operations related
to name collisions.  Names in the resulting set or relation are now
verified individually that they are unique.  This ensures that there are
no ambiguigies with respect to names in the resulting set or relation.
This commit completes the implementations of the operations in the
SparseFormula related classes (all except the 'bounds' method).

Closes #233