Naming fixes, sparse formula simplification, misc
This commit should probably have been split into multiple smaller
commits, but they've all become interrelated, so here is one large set
of changes.  The major changes of this commit are:

Updated tuple variable renaming:
  This update changes the way the initial parsing/processing of
sets/relations when the same tuple variable name is present in the
formula string.

Start of Fourier-Motzkin implementation:
  The sparse formula classes now have the start of simplifications
implemented.  Variables present in equalities can be successfully
projected out in most situations by replacing uses of the variable with
the expression it is equal to.  The traditional FM upper/lower bound
implementation needs to be completed still.

Sparse formula union fix:
  The union operation wasn't properly combining the free variables of
the two formulas being unioned.  This commit fixes this and ensures that
when the formulas are combined, unique variables are created and proper
mappings are made from the old names to the new names.

Removal of _columns and _functions fields of SparseFormula:
  The _columns and _functions fields of the SparseFormula class were
removed as they were unnecessary.  This did cause the add_function
routine to be slightly changed and renamed to get_function.  Also, the
__hash__ implementation in SparseFormula changed to only use the hash of
the formula's disjunction to compute the hash.

References #238,#233