Switched to SparseSet/Relation backend
This commit switches the set/relation backend to use the new data
structure that is much faster.  The SparseSet and SparseRelation classes
were renamed to Set and Relation.  Many tests were updated or disabled.
All tests that are still enabled pass except for the PyCLooG tests.  A
routine still needs to be written to create a constraint matrix that can
be passed to CLooG.  The old Set/Relation classes are not used anymore
and are no longer installed.

References #240