Cleaned up copy() method signatures
This commit simplifies the signatures of the copy() methods for the
various Set/Relation-related classes.  Previously each copy method had
to have a long string of keyword arguments for every keyword used in
every copy method.  This commit adds the use of **kwargs in the
signatures and calls of these methods.  This greatly simplifies the
code.  Only specific keyword arguments that are needed in a particular
copy method are declared explicitly.