Implemented contains_nest and function collapsing
The contains_nest method on the Set/Relation classes return True if that
Set or Relation contains the given function nesting.  Building on this,
the copy methods now accept the 'collapse' argument.  This argument is
used to specify a set of function nestings to be collapsed while making
the copy.  Tests were added to ensure these features are working
properly.