Added another inverse function simplification rule
A new rule was added that again attempts to manipulate equality
constraints to be in terms of later tuple variables.  This is needed
when generating code and we need to ask the question 'what expression is
the given output tuple variable equal to'.  Specifically, the equality
constraint i=f(j) will be converted to j=f_inv(i) if j is a later tuple
variable and f_inv is the inverse of f.  Tests were added for this as
well.