moldyn_fst_cpack.spec example now working!
This commit makes many changes to the code generator to use the new data
structure rather than the old one.  The example spec file
moldyn_fst_cpack.spec now generates valid code.  This shows that the new
data structure is in a good enought state for this branch to be merged
into master.