Fixed inverse simplify bug
This commit fixes a bug related to the signs of coefficients of
arguments to functions.  The inverse function simplification was not
changing the signs of the arguments appropriately.  This was causing
functions nests such as f(-1*g(...))) rather than f(g(...)).  A test was
added to ensure that this bug is fixed.

Closes #245