Fixed equality methods, performance improvements
This commit changes the way equality methods (__eq__) are implemented.
Previously these were just assuming hash(self)==hash(other) was
sufficient for implemented equality.  This was changed to actually tests
equality using the value of the objects being compared.  In addition to
this, a few methods were changed or removed to improve performance.