Updates SparseTileTrans, includes inputs
Adds the data dependences that are input to the sparse tiling routine and the
number of tiles symbolic to the 'inputs' collection.  This was necessary so the
code generation for the IDGCall node included these as input.  Also added
'get_var_name' implementations to the DataDependence and Symbolic classes to
support generation of calls with these types as input.

References #85