Completed FST inspector code generation
Adds the final piece for FST inspector code generation: explicit dependence
generation.  The FST transformation spec file was updated so that code can be
generated for testing as executor codegen was failing.  The calc_equality_value
routine was updated to allow the caller to have the routine ignore inequality
constraints and just consider equality constraints.  The DataDependence class
was updated to include a 'target' field that indicates the target of the
dependence.  Finally, the code for generating explicit depdences was added.

References #85