Fixes a bug related to the SparseTransVisitor
Fixes how variable coefficients are updated when a single variable is present
in a expression multiple times.  For example, a+a.  A defaultdict is used to
start all coefficients off at 0 and the coefficients are simply added up at
each instance.  Before this commit they were being overwritten each time.

Closes #295