Added {Set,Relation}.approximate() implementation
The approximate() method takes a dictionary mapping function names to range
bounds for that function.  The result is a copy of the set or relation with
constraints like tuple_var=f(...) replaced with tuple_var bounded based on the
given bounds for f.  Associated tests were added for this method.

Closes #307