Adds trans_equals() routine
trans_equals() translates a string such as 'a=b' to 'a==b'.  It is robust
enough to avoid translating 'a>=b' to 'a>==b' so as to avoid invalid code.
This routine is used during code generation.  Set/Relation strings use the
single equals syntax to represent equality whereas a single equals in C
represents assignment.  To generate valid code, this routine is needed.