Updates executor codegen, approximate and guards
Modifies executor code generation in three ways:
-Performs approximation on the iteration spaces of all statements before they
are passed to CLooG.  This converts constraints such as t=theta(...) to bound t
based on known range information for theta.
-Removes UFS constraints from the scattering function of each statement before
they are passed to CLooG.  These constraints are not needed for scattering
function purposesand are therefore able to be removed.
-Adds the start of statement guard generation in cases where a UFS constraint
is present in the statement's iteration space.  This is incomplete currently as
the ER API isn't being used to test for equality between Tuple2D structures.

References #85