Updates codegen for ExplicitDependences
The proper variable names are being used now when generating code for
ExplicitDependences.  Issue #320 is still not complete, but after looking at
how to take this approach, I want to avoid it if at all possible.  Determining
if a given variable name is input to some function constraint is messy.  This
commit is the last that is currently needed for FST codegen.  The code being
generated up to FST is verifying.

Closes #85