Updates codegen to specify size info to ER ctors
Updates the code generation routines to specify size information to the
constructors of EF_1D and ER_1DCOO/ExplicitDependence.  This is an optimization
that avoids unnecessary overhead.

Also switches the max/min usage in calc_size_string so it is calculating the
'inner' constraints rather than the 'outer' constraint.

References #368