Enables explicit specification of ER type
Enables a transformation or other code to explicitly define the type of an
Explicit Relation.  This is needed in the spmv CPACK transformation case as
ERG_cpack requires an ER_U1D as input but the spmv computation has only a
single conjunction that needs to be passed in to ERG_cpack.  DataPermuteTrans
now explicitly defines that the input ER has type 'er_u1d'.