Updates codegen of EF_1D explicit functions
Updates the code generation of calls to EF_1D_ctor.  Both the input and output
bounds of the explicit function are required now.  Also, the call must specify
whether or not the EF is a permutation.  Previously the boolean permutation
argument was optional.  This caused issues with the compiler choosing the
incorrect version of EF_1D_ctor.